move most display helpers out
