+++ src/datadec/__init__.py
@@ -5,14 +5,10 @@ warnings.filterwarnings("ignore", message="urllib3 v2 only supports OpenSSL")
 
 from datadec import script_utils  # noqa: E402
 from datadec.data import DataDecide  # noqa: E402
-from datadec.wandb_eval.wandb_store import WandBStore  # noqa: E402
-from datadec.wandb_eval.wandb_downloader import WandBDownloader  # noqa: E402
 from datadec.wandb_eval import analysis_helpers  # noqa: E402
 
 __all__ = [
     "DataDecide",
     "script_utils",
-    "WandBStore",
-    "WandBDownloader",
     "analysis_helpers",
 ]
